security_status_detection_backend/requirements.txt
# Ultralytics requirements
# Usage: pip install -r requirements.txt

# Ultralytics ---------------------------------
ultralytics >= 8.0.81

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python==4.7.0.68
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
# ipython  # interactive notebook
# albumentations>=1.0.3
# pycocotools>=2.0.6  # COCO mAP
# roboflow
certifi>=2022.12.7 # not directly required, pinned by Snyk to avoid a vulnerability


# WebServer ----------------------------------------
Flask
websockets
flask-cors

# Database ----------------------------------------
pymysql>=1.0.3
